Added java code that is used in a Google Cloud blog post.  (#866)
* Renamed cloud-client to java_starter_example

* Removed cloud-client from repo

* Added java_update_cases_blog_code containing final code that is part of a Google Cloud blog post.

* Added java_working_app_example

* Ran java formatter and cleaned code

* Ran formatter and linter

* Made project number as input instead of hard coded

* Changed project id to project number
