+++ examples/cloud-support/java_starter_example/data/case.json
