+++ examples/cloud-support/java_starter_example/pom.xml
