+++ examples/cloud-support/java_starter_example/src/main/java/com/example/gcpsupport/GetCase.java
@@ -37,9 +37,9 @@ public class GetCase {
   public static void main(String[] args) {
 
     try {
-      // TODO(developer): Replace this variable with your project id
-      String projectId = "00000";
-      String PARENT_RESOURCE = String.format("projects/%s", projectId);
+      // TODO(developer): Replace this variable with your project number
+      String projectNumber = "00000";
+      String PARENT_RESOURCE = String.format("projects/%s", projectNumber);
 
       // TODO(developer): Replace this variable with your case id
       String caseId = "00000";
