+++ examples/cloud-support/java_starter_example/src/main/java/com/example/gcpsupport/Starter.java
