+++ examples/cloud-support/java_starter_example/src/test/java/com/example/gcpsupport/GetCaseTest.java
