+++ examples/cloud-support/java_update_cases_blog_code/my-app/README.md
@@ -1,0 +1,6 @@
+# Utilizing Cloud Support API to Programmatically Update Support Cases (Blog Code)
+
+This sample app has the final code written in the Google Cloud Blog post "Utilizing Cloud Support API to Programmatically Update Support Cases." It allows a user to programatically add an email to all open support cases.
+
+
+
