+++ examples/cloud-support/java_working_app_example/my-app/data/case.json
@@ -1,0 +1,15 @@
+{
+    "classification": {
+      "displayName": "Access Transparency > Access Transparency > None",
+      "id": "100H44Q1CDHMASRJ41A74OBEEDO62SJ5DPHNI6GJ85HM6PBJECG58SJ1DPPN0OBICLN66U920H76URJ5"
+    },
+    "displayName": "3rd Test - case - Premium support - Project level case",
+    "description": "This is new sample support case description",
+    "name": "",
+    "state": "NEW",
+    "priority": "P4",
+    "subscriberEmailAddresses": [
+      "test1234@google.com"
+    ],
+    "testCase": true
+  }
